requirements.txt
requests
python-dotenv
argparse